+++ delete.php
@@ -1,23 +1,25 @@
 <?php
+	session_start();
+	if($_SESSION['isEditor']){
+		$filename = $_POST["fileToDelete"];
+		unlink('uploadedimages/'.$filename);
+		unlink('uploadedimages/thumb/'.$filename);
 
-	$filename = $_POST["fileToDelete"];
-	unlink('uploadedimages/'.$filename);
-	unlink('uploadedimages/thumb/'.$filename);
+		$jsonfile = "galleryinfo.json";	
+		$current = file_get_contents($jsonfile);
+		$contents = json_decode($current, true);		
+		foreach ($contents as $index => $content) {		
+			if($content['filename'] == $filename){			
+				unset($contents[$index]);			
+			}		
+		}
+		
+		
+		$fileObj = fopen($jsonfile, "w");
+		fwrite($fileObj, json_encode($contents));
+		fclose($fileObj);
 
-	$jsonfile = "galleryinfo.json";	
-	$current = file_get_contents($jsonfile);
-	$contents = json_decode($current, true);		
-	foreach ($contents as $index => $content) {		
-		if($content['filename'] == $filename){			
-			unset($contents[$index]);			
-		}		
+		//header("Location: index.php"); /* Redirect browser */
+		header('Location: ' . $_SERVER['HTTP_REFERER']);
 	}
-	
-	
-	$fileObj = fopen($jsonfile, "w");
-	fwrite($fileObj, json_encode($contents));
-	fclose($fileObj);
-
-	//header("Location: index.php"); /* Redirect browser */
-	header('Location: ' . $_SERVER['HTTP_REFERER']);
 ?>
